Initial working version (single-node-only)
The test manifests for single-node (everything installed on
the master machine) now work properly.  You have to
edit your puppet.conf by hand to turn on storeconfigs and
point it at puppetdb, but other than that, this successfully
handles everything that you need for either an hsql or
postgres setup.