Add missing `inherit` for `puppetdb::params`
Prior to this commit if you tried to use the class
`puppetdb::master::puppetdb_conf` directly (rather than
implicitly through `puppetdb::master::config`), you would
get a failure about `puppetdb::params` not having been
evaluated.  This is because the class was missing an
`inherits` statement; this commit fixes the bug.