17594 - PuppetDB - Add ability to set standard host listen address and open firewall to standard port
Prior to this commit the module did not provide a way to set a bind address for the HTTP port.  This
commit allows users to not only bind to an address and port other than localhost and 8080, but it also
opens the firewall if explicitly requested.

17594 - PuppetDB - Add ability to set standard host listen address and open firewall to standard port

Addition to intial commit for this issue.  Added comments for the boolean open_http_port paramter.

17594 - PuppetDB - Add ability to set standard host listen address and open firewall to standard port

Addition to intial commit for this issue.  Added comments for the boolean open_http_port paramter.