Add initial rspec-system tests and scaffolding
This adds a basic test to run the standard 1 node setup until it succeeds.

This also fixes a bug with the connection validation facilities. Before this
patch we would keep retrying the connection to PuppetDB until the timeout
expired, this patch fixes that bug making provisioning faster and well, more
reliable. I've also increased the timeout as some systems take longer than
15 seconds for PuppetDB to start.

Signed-off-by: Ken Barber <ken@bob.sh>