Make database_password an optional parameter
In some cases it may not be possible to securely specify a password for
PuppetDB. When that is true, it's impossible to use the module to manage
PuppetDB, because specifying undef for the password will cause database
validation to fail, or for the setting in database.ini to be overridden.

This change will simply check that database_password is set, and will
skip validation and changing database.ini in the case where it is not.