Pin fixtures to use postgresql 2.5.0
Unfortunately postgresql master is not compatible with this module yet, so to
avoid errors during testing we are having to pin the revision of postgresql
we pull down.

Unfortunately again, puppetlabs_spec_helper doesn't yet support branches, so
we're pinning on a specific revision for now.

This can be removed once we are modified to be compatible with the newly
released postgresql module.

Signed-off-by: Ken Barber <ken@bob.sh>