Use the /v2 metrics endpoint instead of /metrics
The base url /metrics has long been a deprecated URL and creates noise when
tested like this. Not to mention in current puppetdb master we are removing
/v1 support so this test starts to break.

By making it /v2, we still afford support for most of our 1.x users, and
since /v2 will be around for quite some time this seems like a suitable
temporary step.

In the future a generic URL should really be provided by PuppetDB as something
that is more permanent and generic.

Signed-off-by: Ken Barber <ken@bob.sh>