Bump version requirement for puppetlabs/stdlib
The is_pe fact was first defined in stdlib in 3.1.0. We want to use that
fact for making the "PE or Not PE" determination in params, so we need
to require a stdlib at least that new. Note that at the time of commit
the current version of stdlib is 4.1.0, so we're not exactly requiring
bleeding edge.