Use str2bool for backwards compatibility, and avoid changing puppetdb_service_status
is_pe being a fact, its smarter her to coerce the value of it too a boolean, in
case we still receive the value as a string. In the future structured facts
will become the norm.

I've also removed the patch against the values for puppetdb_service_status, matching
against a bareword boolean while it looks good, if the value is indeed a boolean the
in will fail in the current parser. This can be brought up again in the future.

Signed-off-by: Ken Barber <ken@bob.sh>