Add logic for manage_pg_repo defaults
The manage_pg_repo defaulted to 'true' in all cases. This proved
to be a problem on supported OSes that weren't of the Redhat or
Debian families as puppetlabs-postgresql only supports managing
Debian or Redhat Postgresql repositories.

The parameter now defaults to true only when $osfamily is Redhat or
Debian