(PDB-4945) Default to json fact cache
Puppetserver 7 now defaults the fact cache to json[1]. When running in
agent-server mode, configure routes.yaml the same way. There is no change in
behavior when running with puppetserver < 7 or `puppet apply`.

As a result of this change, when the puppetdb module is used to install the
puppetdb terminus and routes.yaml, facts will now be cached as non-pretty json
in:

     /opt/puppetlabs/server/data/puppetserver/server_data/facts/<node>.json

Instead of pretty yaml in:

     /opt/puppetlabs/server/data/puppetserver/yaml/<node>.yaml

The difference in pathing is due to the abstract yaml and json termini using
different base directories in which to store data: `Puppet[:yamldir]` vs
`Puppet[:server_data]`.

[1] https://github.com/puppetlabs/puppetserver/blob/794e725628545aa075e3f8b5fa91f21628373594/src/ruby/puppetserver-lib/puppet/server/puppet_config.rb#L56