(SERVER-2500) Use DNF modules to install postgres on RHEL 8
The postgresql module added support for DNF modules to install postgres
at specific versions on RHEL 8. However in order to install a
non-default version, like we need to test Puppet Platform 6, which
only supports Postgres 9.6, `manage_dnf_module` must be set to true for
the `puppetlabs-postgresql` module. This commit ensures that mode is set
when isntalling on RHEL 8.

Note that when the default version is being installed (postgres 10 on
RHEL8), the code path that this flag opts into is already being used.