Set owner of server config.ini to root
The puppetdb user doesn't need to write to the config (it's managed by
Puppet after all) so setting the owner to root and using the group to
only read is safer.

It is also closer to packaging, which ensures the owner is root after a
package update. Ideally packaging would use the same group and mode as
well so there isn't an event after a package update, but this change at
least trims it down from 3 changes to 2.