Archive CHANGELOG.md
When preparing a new release of the module, GCG overrite the whole
CHANGELOG.md file, but it used to be crafted by hand and we do not want
to loose this information.

GCG can append an existing file to the generated changelog, and if this
file is properly formatted, can guess which versions are covered by this
HISTORY file to not include them again.

Move the CHANGELOG.md to HISTORY.md and adjust the release lines so that
GCG can properly parse the first one and the following ones are
formatted the same way.  The resulting CHANGELOG.md which will be
generated on the next release preparation will be consistently
formatted.